+++ spec/renuo_cms_rails/cms_helper_spec.rb
@@ -6,9 +6,10 @@ describe RenuoCmsRails::CmsHelper do
   include RenuoCmsRailsSpecHelper
 
   before(:each) do
-    RenuoCmsRails.configuration.api_host = 'some.host'
-    RenuoCmsRails.configuration.api_key = 'apikey'
-    RenuoCmsRails.configuration.private_api_key = 'pk'
+    RenuoCmsRails.reset
+    RenuoCmsRails.config.api_host = 'some.host'
+    RenuoCmsRails.config.api_key = 'apikey'
+    RenuoCmsRails.config.private_api_key = 'pk'
   end
 
   it 'sets the correct attributes for the public view' do
@@ -17,7 +18,7 @@ describe RenuoCmsRails::CmsHelper do
     node = Capybara.string cms('some.content', 'Some CMS content')
     expect(node).to have_css('div', text: 'Some CMS content')
     expect(node).to have_css("div[data-content-path='some.content-en']")
-    expect(node).to have_css("div[data-api-host='some.host']")
+    expect(node).to have_css("div[data-api-host='https://some.host']")
     expect(node).to have_css("div[data-api-key='apikey']")
     expect(node).not_to have_css('div[data-private-api-key]')
   end
@@ -28,7 +29,7 @@ describe RenuoCmsRails::CmsHelper do
     node = Capybara.string cms('some.content', 'Some CMS content')
     expect(node).to have_css('div', text: 'Some CMS content')
     expect(node).to have_css("div[data-content-path='some.content-en']")
-    expect(node).to have_css("div[data-api-host='some.host']")
+    expect(node).to have_css("div[data-api-host='https://some.host']")
     expect(node).to have_css("div[data-api-key='apikey']")
     expect(node).to have_css("div[data-private-api-key='pk']")
   end
@@ -40,7 +41,7 @@ describe RenuoCmsRails::CmsHelper do
     expect(node).not_to have_css('div', text: 'Some CMS content')
     expect(node).to have_css('div p', text: 'Block content!')
     expect(node).to have_css("div[data-content-path='some.content-en']")
-    expect(node).to have_css("div[data-api-host='some.host']")
+    expect(node).to have_css("div[data-api-host='https://some.host']")
     expect(node).to have_css("div[data-api-key='apikey']")
     expect(node).not_to have_css('div[data-private-api-key]')
   end
@@ -52,8 +53,17 @@ describe RenuoCmsRails::CmsHelper do
     node = Capybara.string cms('some.content')
     expect(node).to have_css('div', text: 'I18n.t content!')
     expect(node).to have_css("div[data-content-path='some.content-en']")
-    expect(node).to have_css("div[data-api-host='some.host']")
+    expect(node).to have_css("div[data-api-host='https://some.host']")
     expect(node).to have_css("div[data-api-key='apikey']")
     expect(node).not_to have_css('div[data-private-api-key]')
+  end
+
+  it 'uses the custom content_path_generator config' do
+    allow(self).to receive(:cms_admin?).and_return(false)
+
+    RenuoCmsRails.config.content_path_generator = ->(path) { "#{I18n.locale}--#{path}" }
+
+    node = Capybara.string cms('some.content', 'Some CMS content')
+    expect(node).to have_css("div[data-content-path='en--some.content']")
   end
 end