+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Version 0.1.0
+
+* add https:// by default to api_host if no protocol is given
+* better dependencies specs
+* make the content path configurable
+* refactoring
+
 ## Version 0.0.1
 
 * initial release